Formated!

+++ src/me/Straiker123/NoteColor.java
@@ -2,6 +2,7 @@ package me.Straiker123;
 
 public class NoteColor {
 	private int note;
+
 	public NoteColor(int note) {
 		if (note < 0) {
 			throw new IllegalArgumentException("The note value is lower than 0");
@@ -10,16 +11,17 @@ public class NoteColor {
 			throw new IllegalArgumentException("The note value is higher than 24");
 		}
 		this.note = note;
-}
+	}
 
-public float getValueX() {
-	return (float) note / 24F;
-}
+	public float getValueX() {
+		return note / 24F;
+	}
 
-public float getValueY() {
-	return 0;
-}
-public float getValueZ() {
-	return 0;
-}
+	public float getValueY() {
+		return 0;
+	}
+
+	public float getValueZ() {
+		return 0;
+	}
 }
